==> 1. Demographics null treatment.ipynb <==
## <center> Null Value treatment for Train_Demographic Using Model Based Imputation

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.preprocessing import OneHotEncoder, LabelEncoder, MinMaxScaler
from sklearn.model_selection import train_test_split
from sklearn.neighbors import KNeighborsClassifier

import joblib

import warnings
warnings.filterwarnings("ignore")

dfd = pd.read_csv(r"C:\Users\Soorya\Documents\capstone-files\TrainData\Data\Train_Demographics.csv",na_values= ['?','MISSEDDATA','MISSINGVALUE','NA'])

# can drop Country as all values are same 
dfd.drop("Country",axis=1,inplace=True)

# seperating/storing the rows with not-null values
df2=dfd.dropna().copy()

df2.head()

df2.shape

# index value for filtering rows with null values
df2_index = set(df2.index.values)
dfd_index = set(dfd.index.values)
df3_index = dfd_index.difference(df2_index)

df3 = dfd.filter(items=list(df3_index), axis=0).copy()

# all rows with null values
df3.head()

df3.shape

dfd_notnull = df2.copy()
dfd_null = df3.copy()

## Perform data preparation 

# numerical encoding the InsuredGender column
df2['InsuredGender'] = df2.InsuredGender.astype('category').cat.codes

#onehot encoding

# Converting type of columns to category
df2['InsuredEducationLevel'] = df2['InsuredEducationLevel'].astype('category')
df2['InsuredOccupation'] = df2['InsuredOccupation'].astype('category')
df2['InsuredHobbies'] = df2['InsuredHobbies'].astype('category')

# Create an instance of One-hot-encoder
enc = OneHotEncoder(handle_unknown = 'ignore')
enc.fit(df2[["InsuredEducationLevel","InsuredOccupation","InsuredHobbies"]])
enc_data = pd.DataFrame(enc.transform(df2[["InsuredEducationLevel","InsuredOccupation","InsuredHobbies"]]).toarray())

dfd_enc = df2.join(enc_data)
dfd_enc.drop("CustomerID",axis=1,inplace=True)

dfd_enc.drop(["InsuredEducationLevel","InsuredOccupation","InsuredHobbies"],axis=1,inplace=True)

dfd_enc.shape

dfd_enc.head()

dfd_enc.isnull().sum().sum()

dfd_enc.dropna(inplace=True)

x = dfd_enc.drop('InsuredGender',axis=1)
y = dfd_enc['InsuredGender']

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=0.2, random_state=42)

### Training a simple model for missing value imputation

k = list(range(1,30,2))

knn_acc = []

for i in k:
    neigh = KNeighborsClassifier(n_neighbors=i)
    neigh.fit(x_train,y_train)
    knn_acc.append(neigh.score(x_test, y_test))

print("The optimal value for k is",k[np.argmax(np.array(knn_acc))],"with accuracy score",np.max(np.array(knn_acc)))

# Training with the optimal value for k
neigh = KNeighborsClassifier(n_neighbors=k[np.argmax(np.array(knn_acc))])
neigh.fit(x_train,y_train)

joblib.dump(neigh, 'Demographic.joblib')

## to predict

# changes for the dataframe with null values
df3.drop(["InsuredGender","CustomerID"],axis=1,inplace=True)

df3['ind'] = np.array([x for x in range(0,30)])
originalindex = list(df3.index.values)

df3.set_index('ind',inplace=True)

df3.head()

# Converting type of columns to category
df3['InsuredEducationLevel'] = df3['InsuredEducationLevel'].astype('category')
df3['InsuredOccupation'] = df3['InsuredOccupation'].astype('category')
df3['InsuredHobbies'] = df3['InsuredHobbies'].astype('category')

enc_array = enc.transform(df3[["InsuredEducationLevel","InsuredOccupation","InsuredHobbies"]]).toarray()

enc_data = pd.DataFrame(enc_array)

dfd_test = df3.join(enc_data)
dfd_test.drop(["InsuredEducationLevel","InsuredOccupation","InsuredHobbies"],axis=1,inplace=True)

dfd_test.head()

dfd_test.shape

pred = neigh.predict(dfd_test)

pred

#adding predictions to dfd_null dataframe
dfd_null['InsuredGender'] = pred

dfd_null['InsuredGender'] = dfd_null.InsuredGender.apply(lambda x: "MALE" if x else "FEMALE")

dfd_null.head()

# appending the 2 dataframe to get the imputed null values for demographics data
dfd_final = dfd_notnull.append(dfd_null)

dfd_final.shape

dfd_final.head()

dfd_final['InsuredGender'].value_counts()

#save the file as Demographics.csv
dfd_final.to_csv("Demographics.csv")
