==> 2. Policy Claim Null values Processing.ipynb <==
## Null Value treatment for Train_Claim Using Model Based Imputation

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.preprocessing import LabelEncoder, OneHotEncoder
from sklearn.model_selection import train_test_split
from sklearn.neighbors import KNeighborsClassifier
from sklearn.linear_model import LogisticRegression

import joblib

import warnings
warnings.filterwarnings("ignore")

dfc = pd.read_csv(r"C:\Users\Soorya\Documents\capstone-files\TrainData\Data\Train_Claim.csv",na_values= [ "?", "-5", "MISSINGVALUE", "MISSEDDATA"])
print(dfc.shape)
dfc.head()

dfc.isnull().sum()

# -1 for No witness info in the dataset
dfc['Witnesses'].fillna(-1,axis=0,inplace=True)

# median imputation for the AmountOfTotalClaim
dfc['AmountOfTotalClaim'].fillna(dfc['AmountOfTotalClaim'].median(),axis=0,inplace=True)

# null value is replaced with "NoReport"
dfc['PoliceReport'].fillna("NoReport",axis=0,inplace=True)
dfc['PoliceReport'].value_counts()

# null value is replaced with "Missing collision"
dfc['TypeOfCollission'].fillna("Missing Info",axis=0,inplace=True)

dfc['TypeOfCollission'].value_counts()

# null value is replaced with -10.0
dfc['IncidentTime'].fillna(-10.0,axis=0,inplace=True)

dfc["Year"] = dfc['DateOfIncident'].apply(lambda x: int(x.split("-")[0]))
dfc["Month"] = dfc['DateOfIncident'].apply(lambda x: int(x.split("-")[1]))
dfc["days"] = dfc['DateOfIncident'].apply(lambda x: int(x.split("-")[2]))

dfc.drop("DateOfIncident",axis=1,inplace=True)

dfc.isnull().sum().sum()

## Model Based Null value Imputation for PropertyDamage column

#Before building a model lets perform data pre-processing

# seperating the rows with not-null values
df2 = dfc.dropna().copy()

print(df2.shape)
df2.head()

# index value for filtering rows with null values
df2_index = set(df2.index.values)
dfc_index = set(dfc.index.values)
df3_index = dfc_index.difference(df2_index)

df3 = dfc.filter(items=list(df3_index), axis=0).copy()

# all rows with null values
df3.head()

df3.shape

dfc_notnull = df2.copy()
dfc_null = df3.copy()

## Perform data preparation 

df2['PropertyDamage'].head()

le = LabelEncoder()
df2['CustomerID'] = le.fit_transform(df2['CustomerID'])
df2['IncidentAddress'] = le.fit_transform(df2['IncidentAddress'])

df3['CustomerID'] = le.fit_transform(df3['CustomerID'])
df3['IncidentAddress'] = le.fit_transform(df3['IncidentAddress'])

for i in  df2.columns:
    if df2[i].dtype == 'object':
        df2[i] = df2[i].astype("category").cat.codes
        
df3.drop(["PropertyDamage"],axis=1,inplace=True)
for i in  df3.columns:
    if df3[i].dtype == 'object':
        df3[i] = df3[i].astype("category").cat.codes

x = df2.drop('PropertyDamage',axis=1)
y = df2['PropertyDamage']

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=0.2, random_state=42)

k = list(range(1,30,2))

knn_acc = []

for i in k:
    neigh = KNeighborsClassifier(n_neighbors=i)
    neigh.fit(x_train,y_train)
    knn_acc.append(neigh.score(x_test, y_test))

print("The optimal value for k is",k[np.argmax(np.array(knn_acc))],"with accuracy score",np.max(np.array(knn_acc)))

k = [10**x for x in range(-4,5)]

lr_acc = []

for i in k:
    lr = LogisticRegression(C=i)
    lr.fit(x_train,y_train)
    lr_acc.append(lr.score(x_test, y_test))

print("The optimal value for k is",k[np.argmax(np.array(lr_acc))],"with accuracy score",np.max(np.array(lr_acc)))

# Training KNN with optimal K=1 on entire dataset 
neigh = KNeighborsClassifier(n_neighbors=1)
neigh.fit(x,y)

joblib.dump(neigh, 'policyClaim.joblib')

## predict for null values 

pred = neigh.predict(df3)
pred

#adding predictions to dfd_null dataframe
dfc_null['PropertyDamage'] = pred

dfc_null['PropertyDamage'] = dfc_null.PropertyDamage.apply(lambda x: "YES" if x else "NO")

dfc_null.head()

dfc_null['PropertyDamage'].value_counts()

# appending the 2 dataframe to get the imputed null values for demographics data
dfc_final = dfc_notnull.append(dfc_null)

dfc_final.shape

dfc_final.head()

dfc_final.to_csv("PolicyClaim.csv",index=False)
