==> 3. Vehicle's data processing.ipynb <==
import pandas as pd
import numpy as np


dfv = pd.read_csv(r"C:\Users\Soorya\Documents\capstone-files\TrainData\Data\Train_Vehicle.csv",na_values=['???'])
print(dfv.shape)
dfv.head()

def YOM(x):
    a = dfv[dfv['CustomerID']==x].copy()
    return int(a[a['VehicleAttribute']=="VehicleYOM"]["VehicleAttributeDetails"].values)

dfv['VehicleYOM'] = dfv['CustomerID'].apply(lambda x: YOM(x))

def Model(x):
    a = dfv[dfv['CustomerID']==x].copy()
    return a[a['VehicleAttribute']=="VehicleModel"]["VehicleAttributeDetails"].values

dfv['VehicleModel'] = dfv['CustomerID'].apply(lambda x: Model(x))

def ID(x):
    a = dfv[dfv['CustomerID']==x].copy()
    return str(a[a['VehicleAttribute']=="VehicleID"]["VehicleAttributeDetails"].values)

dfv['VehicleID'] = dfv['CustomerID'].apply(lambda x: ID(x))

def Make(x):
    a = dfv[dfv['CustomerID']==x].copy()
    return a[a['VehicleAttribute']=="VehicleMake"]["VehicleAttributeDetails"].values

dfv['VehicleMake'] = dfv['CustomerID'].apply(lambda x: Make(x))

dfv.head()

dfv.drop(["VehicleAttribute","VehicleAttributeDetails"],axis=1,inplace=True)

dfv.head()

dfv.to_csv("Vehicles_clean.csv",index=False)

ff = pd.read_csv("Vehicles_clean.csv")
ff.isnull().sum().sum()
