==> 4. Data Merging and Visualization -Original.ipynb <==
# <center>Fraud Detection

### The data provided is of all the customers who's insurnce is been claimed and in that customers whom claim is classified asfraud or not.

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

from sklearn.preprocessing import OneHotEncoder

## Reading the Demographics data set (after removing null values)

# Demographics Data : These files consist of the demographic data of each customer
dfd = pd.read_csv("Demographics.csv",na_values= ["", "NA"])
dfd.drop("Unnamed: 0",axis=1,inplace=True)
print(dfd.shape)
dfd.head()

# Missing values in the Demographics data set
dfd.isnull().sum().sum()

## Reading the policy claim data set

# Data of Claim : These files consist of the details about the insurance claim,that the customer applied for.
dfc = pd.read_csv("PolicyClaim.csv",na_values= [ "?", "-5", "MISSINGVALUE", "MISSEDDATA"])
print(dfc.shape)
dfc.head()

# index 16 and 17 is fishy -- check if they are fraud case

# Missing values in the policy claim dataframe
dfc.isnull().sum().sum()

### Here maybe the missing value itself maybe an information, so for some columns i have imputed a new value that indicates that value was missing and for property damage i have done model based imputation using KNN.

## Reading the Policy Information data set

# Policy Information: insurance policy information, connected to the claim with the insurance company.
dfp = pd.read_csv(r"C:\Users\Soorya\Documents\capstone-files\TrainData\Data\Train_Policy.csv",na_values=["-1", "MISSINGVAL","NA"])
print(dfp.shape)
dfp.head()

#Policy_CombinedSingleLimit -- seems to convert into percentage 
#InsuredRelationship -- maybe marital status

dfp.isnull().sum().sum()

dfp['PolicyAnnualPremium'].median()

# null value is replaced with median value

dfp['PolicyAnnualPremium'].fillna(dfp['PolicyAnnualPremium'].median(),inplace=True)

dfp.isnull().sum().sum()

## Reading the Vehicle Information data set

#Data of Vehicle : These files consist of the details about the Vehicle, connected to the policy. 
dfv = pd.read_csv("Vehicles_clean.csv",na_values=["???"])
dfv['VehicleModel'] = dfv['VehicleModel'].apply(lambda x: x[2:-2])
dfv['VehicleID'] = dfv['VehicleID'].apply(lambda x: x[2:-2])
dfv['VehicleMake'] = dfv['VehicleMake'].apply(lambda x: x[2:-2])
print(dfv.shape)
dfv.head()

dfv.isnull().sum().sum()

dfv.duplicated().sum()

dfv.drop_duplicates(inplace=True)

dfv.shape

## Reading the Fraud Information data set

# Fraud Data : This Train.csv contains the Fraud information details
dfwt = pd.read_csv(r"C:\Users\Soorya\Documents\capstone-files\TrainData\Data\Traindata_with_Target.csv")
print(dfwt.shape)
dfwt.head()

dfwt.isnull().sum().sum()

print(dfc.shape)
print(dfd.shape)
print(dfp.shape)
print(dfv.shape)
print(dfwt.shape)

# we need to perform merge on all the dataframes.
df = dfd.merge(dfp).merge(dfv).merge(dfc).merge(dfwt)

df.shape

df.head()

df.isnull().sum().sum()

# storing the file merged file as final_data.csv
df.to_csv("final_data.csv")
